Rename filestate to DIY
