+++ changelog/config.yaml
@@ -4,7 +4,7 @@ types:
   chore: "Miscellaneous"
 scopes:
   auto: [dotnet, go, java, nodejs, python, yaml]
-  backend: [filestate, service]
+  backend: [diy, service]
   build: []
   ci: []
   cli: [about, config, display, engine, import, new, plugin, package, state]
