+++ pkg/backend/diy/stack.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package filestate
+package diy
 
 import (
 	"context"
@@ -34,28 +34,28 @@ import (
 	"github.com/pulumi/pulumi/sdk/v3/go/common/util/result"
 )
 
-// localStack is a local stack descriptor.
-type localStack struct {
+// diyStack is a diy stack descriptor.
+type diyStack struct {
 	// the stack's reference (qualified name).
-	ref *localBackendReference
+	ref *diyBackendReference
 	// a snapshot representing the latest deployment state, allocated on first use. It's valid for the
 	// snapshot itself to be nil.
 	snapshot atomic.Pointer[*deploy.Snapshot]
 	// a pointer to the backend this stack belongs to.
-	b *localBackend
+	b *diyBackend
 }
 
-func newStack(ref *localBackendReference, b *localBackend) backend.Stack {
+func newStack(ref *diyBackendReference, b *diyBackend) backend.Stack {
 	contract.Requiref(ref != nil, "ref", "ref was nil")
 
-	return &localStack{
+	return &diyStack{
 		ref: ref,
 		b:   b,
 	}
 }
 
-func (s *localStack) Ref() backend.StackReference { return s.ref }
-func (s *localStack) Snapshot(ctx context.Context, secretsProvider secrets.Provider) (*deploy.Snapshot, error) {
+func (s *diyStack) Ref() backend.StackReference { return s.ref }
+func (s *diyStack) Snapshot(ctx context.Context, secretsProvider secrets.Provider) (*deploy.Snapshot, error) {
 	if v := s.snapshot.Load(); v != nil {
 		return *v, nil
 	}
@@ -68,78 +68,78 @@ func (s *localStack) Snapshot(ctx context.Context, secretsProvider secrets.Provi
 	s.snapshot.Store(&snap)
 	return snap, nil
 }
-func (s *localStack) Backend() backend.Backend              { return s.b }
-func (s *localStack) Tags() map[apitype.StackTagName]string { return nil }
+func (s *diyStack) Backend() backend.Backend              { return s.b }
+func (s *diyStack) Tags() map[apitype.StackTagName]string { return nil }
 
-func (s *localStack) Remove(ctx context.Context, force bool) (bool, error) {
+func (s *diyStack) Remove(ctx context.Context, force bool) (bool, error) {
 	return backend.RemoveStack(ctx, s, force)
 }
 
-func (s *localStack) Rename(ctx context.Context, newName tokens.QName) (backend.StackReference, error) {
+func (s *diyStack) Rename(ctx context.Context, newName tokens.QName) (backend.StackReference, error) {
 	return backend.RenameStack(ctx, s, newName)
 }
 
-func (s *localStack) Preview(
+func (s *diyStack) Preview(
 	ctx context.Context,
 	op backend.UpdateOperation, events chan<- engine.Event,
 ) (*deploy.Plan, display.ResourceChanges, result.Result) {
 	return backend.PreviewStack(ctx, s, op, events)
 }
 
-func (s *localStack) Update(ctx context.Context, op backend.UpdateOperation) (display.ResourceChanges, result.Result) {
+func (s *diyStack) Update(ctx context.Context, op backend.UpdateOperation) (display.ResourceChanges, result.Result) {
 	return backend.UpdateStack(ctx, s, op)
 }
 
-func (s *localStack) Import(ctx context.Context, op backend.UpdateOperation,
+func (s *diyStack) Import(ctx context.Context, op backend.UpdateOperation,
 	imports []deploy.Import,
 ) (display.ResourceChanges, result.Result) {
 	return backend.ImportStack(ctx, s, op, imports)
 }
 
-func (s *localStack) Refresh(ctx context.Context, op backend.UpdateOperation) (display.ResourceChanges, result.Result) {
+func (s *diyStack) Refresh(ctx context.Context, op backend.UpdateOperation) (display.ResourceChanges, result.Result) {
 	return backend.RefreshStack(ctx, s, op)
 }
 
-func (s *localStack) Destroy(ctx context.Context, op backend.UpdateOperation) (display.ResourceChanges, result.Result) {
+func (s *diyStack) Destroy(ctx context.Context, op backend.UpdateOperation) (display.ResourceChanges, result.Result) {
 	return backend.DestroyStack(ctx, s, op)
 }
 
-func (s *localStack) Watch(ctx context.Context, op backend.UpdateOperation, paths []string) result.Result {
+func (s *diyStack) Watch(ctx context.Context, op backend.UpdateOperation, paths []string) result.Result {
 	return backend.WatchStack(ctx, s, op, paths)
 }
 
-func (s *localStack) GetLogs(ctx context.Context, secretsProvider secrets.Provider, cfg backend.StackConfiguration,
+func (s *diyStack) GetLogs(ctx context.Context, secretsProvider secrets.Provider, cfg backend.StackConfiguration,
 	query operations.LogQuery,
 ) ([]operations.LogEntry, error) {
 	return backend.GetStackLogs(ctx, secretsProvider, s, cfg, query)
 }
 
-func (s *localStack) ExportDeployment(ctx context.Context) (*apitype.UntypedDeployment, error) {
+func (s *diyStack) ExportDeployment(ctx context.Context) (*apitype.UntypedDeployment, error) {
 	return backend.ExportStackDeployment(ctx, s)
 }
 
-func (s *localStack) ImportDeployment(ctx context.Context, deployment *apitype.UntypedDeployment) error {
+func (s *diyStack) ImportDeployment(ctx context.Context, deployment *apitype.UntypedDeployment) error {
 	return backend.ImportStackDeployment(ctx, s, deployment)
 }
 
-func (s *localStack) DefaultSecretManager(info *workspace.ProjectStack) (secrets.Manager, error) {
+func (s *diyStack) DefaultSecretManager(info *workspace.ProjectStack) (secrets.Manager, error) {
 	return passphrase.NewPromptingPassphraseSecretsManager(info, false /* rotatePassphraseSecretsProvider */)
 }
 
-type localStackSummary struct {
+type diyStackSummary struct {
 	name backend.StackReference
 	chk  *apitype.CheckpointV3
 }
 
-func newLocalStackSummary(name backend.StackReference, chk *apitype.CheckpointV3) localStackSummary {
-	return localStackSummary{name: name, chk: chk}
+func newDIYStackSummary(name backend.StackReference, chk *apitype.CheckpointV3) diyStackSummary {
+	return diyStackSummary{name: name, chk: chk}
 }
 
-func (lss localStackSummary) Name() backend.StackReference {
+func (lss diyStackSummary) Name() backend.StackReference {
 	return lss.name
 }
 
-func (lss localStackSummary) LastUpdate() *time.Time {
+func (lss diyStackSummary) LastUpdate() *time.Time {
 	if lss.chk != nil && lss.chk.Latest != nil {
 		if t := lss.chk.Latest.Manifest.Time; !t.IsZero() {
 			return &t
@@ -148,7 +148,7 @@ func (lss localStackSummary) LastUpdate() *time.Time {
 	return nil
 }
 
-func (lss localStackSummary) ResourceCount() *int {
+func (lss diyStackSummary) ResourceCount() *int {
 	if lss.chk != nil && lss.chk.Latest != nil {
 		count := len(lss.chk.Latest.Resources)
 		return &count
