+++ sdk/go/auto/errors.go
@@ -48,8 +48,8 @@ func IsConcurrentUpdateError(e error) bool {
 	}
 
 	conflictText := "[409] Conflict: Another update is currently in progress."
-	localBackendConflictText := "the stack is currently locked by"
-	return strings.Contains(ae.stderr, conflictText) || strings.Contains(ae.stderr, localBackendConflictText)
+	diyBackendConflictText := "the stack is currently locked by"
+	return strings.Contains(ae.stderr, conflictText) || strings.Contains(ae.stderr, diyBackendConflictText)
 }
 
 // IsSelectStack404Error returns true if the error was a result of selecting a stack that does not exist.
