+++ sdk/go/auto/remote_stack.go
@@ -146,7 +146,7 @@ func (s *RemoteStack) History(ctx context.Context, pageSize, page int) ([]Update
 // Cancel stops a stack's currently running update. It returns an error if no update is currently running.
 // Note that this operation is _very dangerous_, and may leave the stack in an inconsistent state
 // if a resource operation was pending when the update was canceled.
-// This command is not supported for local backends.
+// This command is not supported for diy backends.
 func (s *RemoteStack) Cancel(ctx context.Context) error {
 	return s.stack.Cancel(ctx)
 }
