+++ sdk/python/lib/pulumi/automation/_remote_stack.py
@@ -133,7 +133,7 @@ class RemoteStack:
         Cancel stops a stack's currently running update. It returns an error if no update is currently running.
         Note that this operation is _very dangerous_, and may leave the stack in an inconsistent state
         if a resource operation was pending when the update was canceled.
-        This command is not supported for local backends.
+        This command is not supported for diy backends.
         """
         self.__stack.cancel()
