+++ sdk/python/lib/pulumi/policy/__main__.py
@@ -50,7 +50,7 @@ def main():
                 project=os.environ["PULUMI_PROJECT"],
                 stack=os.environ["PULUMI_STACK"],
                 dry_run=os.environ["PULUMI_DRY_RUN"] == "true",
-                # PULUMI_ORGANIZATION might not be set for filestate backends
+                # PULUMI_ORGANIZATION might not be set for diy backends
                 organization=os.environ.get("PULUMI_ORGANIZATION", "organization"),
             )
         )
