+++ tests/integration/backend/diy/backend_azure_test.go
@@ -1,4 +1,4 @@
-package filestate
+package diy
 
 import (
 	"os"
