+++ tests/integration/backend/diy/backend_gcp_test.go
@@ -1,4 +1,4 @@
-package filestate
+package diy
 
 import (
 	"os"
