+++ tests/integration/backend/diy/helpers.go
@@ -1,4 +1,4 @@
-package filestate
+package diy
 
 import (
 	"crypto/rand"
