+++ tests/integration/backend/diy/project/Pulumi.yaml
